disable LTCG for material factory
Something in here with shader resource names breaks on the newest VS version when LTCG is used. That might mean their is UB lurking in the file or in a library it uses. But I don not have time to investigate further at the moment, disabling LTCG for the file resolves the issue for now.

It could also just be the compiler has a bug.